Add initial sowing functionality

+++ app/routes/crops/$cropId.tsx
@@ -4,13 +4,13 @@ import { json } from "@remix-run/node";
 import { Form, useCatch, useLoaderData } from "@remix-run/react";
 import invariant from "tiny-invariant";
 
-import type { Crop } from "~/models/crops.server";
+import { createSowing, CropWithSowings } from "~/models/crops.server";
 import { deleteCrop } from "~/models/crops.server";
 import { getCrop } from "~/models/crops.server";
 import { requireUserId } from "~/session.server";
 
 type LoaderData = {
-  crop: Crop;
+  crop: CropWithSowings;
 };
 
 export const loader: LoaderFunction = async ({ request, params }) => {
@@ -30,9 +30,24 @@ export const action: ActionFunction = async ({ request, params }) => {
   const userId = await requireUserId(request);
   invariant(params.cropId, "cropId not found");
 
-  await deleteCrop({ userId, id: params.cropId });
+  const formData = await request.formData();
 
-  return redirect("/crops");
+  if (formData.get("action") === "delete") {
+    await deleteCrop({ userId, id: params.cropId });
+    return redirect("/crops");
+  }
+
+  if (formData.get("action") === "add-sowing") {
+    const crop = await getCrop({ userId, id: params.cropId });
+
+    if (!crop) {
+      throw new Response("Not Found", { status: 404 });
+    }
+
+    await createSowing({ id: crop.id });
+
+    return null;
+  }
 };
 
 export default function CropDetailsPage() {
@@ -41,12 +56,31 @@ export default function CropDetailsPage() {
   return (
     <div>
       <h3 className="text-2xl font-bold">{data.crop.species}</h3>
+      <h4 className="text-xl">Sowings</h4>
+      <Form method="post">
+        <button
+          type="submit"
+          name="action"
+          value="add-sowing"
+          className="rounded bg-blue-500 py-2 px-4 text-xs text-white hover:bg-blue-600 focus:bg-blue-400"
+        >
+          Add sowing
+        </button>
+      </Form>
+      {data.crop.sowings.length === 0 ? (
+        <span>No sowings yet</span>
+      ) : (
+        <span>{data.crop.sowings.length} sowing(s)</span>
+      )}
+
       <p>Created {new Date(data.crop.createdAt).toLocaleDateString()}</p>
       <p>Updated {new Date(data.crop.updatedAt).toLocaleDateString()}</p>
       <hr className="my-4" />
       <Form method="post">
         <button
           type="submit"
+          name="action"
+          value="delete"
           className="rounded bg-blue-500 py-2 px-4 text-white hover:bg-blue-600 focus:bg-blue-400"
         >
           Delete
